导出增加新需求

+++ src/main/java/com/fairyland/xrobot/modular/xrobot/domain/TaskDevicesExample.java
@@ -805,7 +805,7 @@ public class TaskDevicesExample {
             return (Criteria) this;
         }
 
-        public Criteria andStateEqualTo(Byte value) {
+        public Criteria andStateEqualTo(Integer value) {
             addCriterion("state =", value, "state");
             return (Criteria) this;
         }
@@ -835,7 +835,7 @@ public class TaskDevicesExample {
             return (Criteria) this;
         }
 
-        public Criteria andStateIn(List<Byte> values) {
+        public Criteria andStateIn(List<Integer> values) {
             addCriterion("state in", values, "state");
             return (Criteria) this;
         }
